data_server/requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
contextlib2==21.6.0
Django==3.2.8
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-extensions==3.1.3
djangorestframework==3.12.4
easydict==1.9
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
mysqlclient==2.0.3
oauth2client==4.1.3
protobuf==3.19.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
rsa==4.8
schedule==1.1.0
schema==0.7.4
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
waiting==1.4.1
